go over all error messages
