+++ goto/commands/add.py
@@ -1,4 +1,4 @@
-from ..gotomagic.text import Error
+from ..gotomagic.text import GotoError
 
 def add(magic, args):
     """
@@ -6,15 +6,15 @@ def add(magic, args):
     """
 
     if (len(args) == 0):
-        return None, Error("missing_magicword_and_uri")
+        return None, GotoError("missing_magicword_and_uri")
 
     if (len(args) == 1):
-        return None, Error("missing_uri")
+        return None, GotoError("missing_uri", magicword=args[0])
 
-    word = args[0]
+    magicword = args[0]
     uri = args[1]
 
-    magic.add_shortcut(word, uri)
+    magic.add_shortcut(magicword, uri)
     magic.save()
 
-    return 'Added magic word %s' % word, None
+    return 'Added magic word %s' % magicword, None
