+++ goto/commands/copy.py
@@ -1,5 +1,5 @@
 import pyperclip
-from ..gotomagic.text import Error
+from ..gotomagic.text import GotoError
 
 
 def copy(magic, args):
@@ -8,7 +8,7 @@ def copy(magic, args):
     """
 
     if (len(args) == 0):
-        return None, Error("missing_magicword")
+        return None, GotoError("missing_magicword")
 
     word = args[0]
     url = str(magic.get_uri(word))
