+++ goto/commands/subl.py
@@ -1,5 +1,5 @@
 import subprocess
-from ..gotomagic.text import Error
+from ..gotomagic.text import GotoError
 
 
 def subl(magic, _):
@@ -9,12 +9,12 @@ def subl(magic, _):
 
     try:
         code = magic['code']
-    except KeyError:
-        return None, Error("no_magicword_named_code")
+    except KeyGotoError:
+        return None, GotoError("no_magicword_named_code")
 
     try:
         subprocess.check_call('subl "%s"' % code, shell=True)
-    except subprocess.CalledProcessError:
-        return None, Error("subl_launch_failed")
+    except subprocess.CalledProcessGotoError:
+        return None, GotoError("subl_launch_failed")
 
     return None, None
