+++ goto/commands/vscode.py
@@ -1,5 +1,5 @@
 import subprocess
-from ..gotomagic.text import Error
+from ..gotomagic.text import GotoError
 
 
 def vscode(magic, _):
@@ -10,11 +10,11 @@ def vscode(magic, _):
     try:
         code = magic['code']
     except KeyError:
-        return None, Error("no_magicword_named_code")
+        return None, GotoError("no_magicword_named_code")
 
     try:
         subprocess.check_call('code "%s"' % code, shell=True)
     except subprocess.CalledProcessError:
-        return None, Error("vscode_launch_failed")
+        return None, GotoError("vscode_launch_failed")
 
     return None, None
