+++ goto/gotomagic/text/__init__.py
@@ -1,0 +1,5 @@
+# -*- code: utf-8 -*-
+"""
+    Text used by GOTO to do UX.
+"""
+from .text import GotoError, GotoWarning, print_text
